Add clippy and fmt to CI
As part of the commit, the use of cargo-fmt as well as cargo-clippy is being added to CI. In addition, the specification of rust-toolchain as stable is added, which is a good thing, since it can be difficult to debug this on the basis of an arbitrary toolchain (e.g. when new checks are added to clippy).